Move bytegen_speedtest.py in neubot/speedtest
