+++ neubot/speedtest/bytegen.py
